refactor : 핸들러 적용을 위한 코드 변경
핸들러들을 따로 구현하면서 기존 코드 그에 맞게 변경
